@@ -10,7 +10,8 @@
   "dependencies": {
     "bootstrap": "^5.3.0",
     "core-js": "^3.8.3",
-    "vue": "^2.6.14"
+    "vue": "^3.0.2",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
