@@ -1,10 +1,10 @@
-import Vue from 'vue'
-import App from './App.vue'
+import { createApp } from 'vue';
+import App from './App.vue';
 import 'bootstrap/dist/css/bootstrap.css';
 import 'bootstrap';
+import store from './store';
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+const app = createApp(App);
+app.config.productionTip = false;
+app.use(store);
+app.mount('#app');
