@@ -1,31 +1,30 @@
 <template>
   <div>
     <FormularioDatos @submit="addData" />
-    <TablaReactiva :data="data" @remove="removeData" />
+    <TablaReactiva :formData="formData" @remove="removeData" />
   </div>
 </template>
 
 <script>
-import FormularioDatos from './components/FormularioDatos.vue';
-import TablaReactiva from './components/TablaReactiva.vue';
+import FormularioDatos from "./components/FormularioDatos.vue";
+import TablaReactiva from "./components/TablaReactiva.vue";
+import { mapState } from "vuex";
 
 export default {
   components: {
     FormularioDatos,
-    TablaReactiva
+    TablaReactiva,
   },
-  data() {
-    return {
-      data: []
-    };
+  computed: {
+    ...mapState(["formData"]),
   },
   methods: {
     addData(newData) {
-      this.data.push(newData);
+      this.formData.push(newData);
     },
     removeData(index) {
-      this.data.splice(index, 1);
-    }
-  }
+      this.formData.splice(index, 1);
+    },
+  },
 };
 </script>
