@@ -4,19 +4,47 @@
     <form @submit.prevent="submitForm" class="row g-3" ref="form">
       <div class="col-md-6">
         <label for="name" class="form-label">Nombre:</label>
-        <input type="text" id="name" v-model="name" class="form-control" minlength="3" maxlength="50" required>
+        <input
+          type="text"
+          id="name"
+          v-model="name"
+          class="form-control"
+          minlength="3"
+          maxlength="50"
+          required
+        />
       </div>
       <div class="col-md-6">
         <label for="email" class="form-label">Email:</label>
-        <input type="email" id="email" v-model="email" class="form-control" required>
+        <input
+          type="email"
+          id="email"
+          v-model="email"
+          class="form-control"
+          required
+        />
       </div>
       <div class="col-md-6">
         <label for="age" class="form-label">Edad:</label>
-        <input type="number" id="age" v-model.number="age" class="form-control" min="18" required>
+        <input
+          type="number"
+          id="age"
+          v-model.number="age"
+          class="form-control"
+          min="18"
+          required
+        />
       </div>
       <div class="col-md-6">
         <label for="phone" class="form-label">Teléfono:</label>
-        <input type="tel" id="phone" v-model="phone" class="form-control" pattern="[0-9\-]+" required>
+        <input
+          type="tel"
+          id="phone"
+          v-model="phone"
+          class="form-control"
+          pattern="[0-9\-]+"
+          required
+        />
       </div>
       <div class="col-12">
         <button type="submit" class="btn btn-primary">Agregar</button>
@@ -26,38 +54,47 @@
 </template>
 
 <script>
+import { mapActions } from "vuex";
+
 export default {
   data() {
     return {
-      name: '',
-      email: '',
+      name: "",
+      email: "",
       age: 0,
-      phone: ''
+      phone: "",
     };
   },
   methods: {
+    ...mapActions(["submitFormData"]),
     submitForm() {
       if (this.$refs.form.checkValidity() && this.validateAge()) {
-        const newData = {
+        const formData = {
           name: this.name,
           email: this.email,
           age: this.age,
-          phone: this.phone
+          phone: this.phone,
         };
-        this.$emit('submit', newData);
-        this.name = '';
-        this.email = '';
-        this.age = 0;
-        this.phone = '';
+        this.submitFormData(formData)
+          .then(() => {
+            this.name = "";
+            this.email = "";
+            this.age = 0;
+            this.phone = "";
+            alert("Usuario creado exitosamente");
+          })
+          .catch((error) => {
+            console.error(error);
+          });
       }
     },
     validateAge() {
       if (this.age < 18) {
-        alert('Debes ser mayor de 18 años para completar el formulario.');
+        alert("Debes ser mayor de 18 años para completar el formulario.");
         return false;
       }
       return true;
-    }
-  }
+    },
+  },
 };
 </script>
